script_plantset/requirements.txt
# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# TPU Support
torch_xla>=2.0.0

# Transformers & NLP
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Computer Vision
Pillow>=9.0.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Text Evaluation
nltk>=3.8.0
rouge-score>=0.1.2

# Progress & Logging
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0

# Utilities
pathlib2>=2.3.7
click>=8.1.0
pyyaml>=6.0.0
requests>=2.31.0

# Optional: Advanced Models
timm>=0.9.0  # For EfficientNet, ViT
accelerate>=0.20.0  # For distributed training
deepspeed>=0.9.0  # For large model training

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0








